refactor(chat): json-encode sse chunks, add timer cleanup, and modularize chat components
